adding MerkeleTree
This commit adds the class MerkleTree to the project. The MerkleTree class is used to create a Merkle tree from an array of transactions and generate a root hash for the tree. The root hash is then used to represent the entire set of transactions in the tree in a more efficient and secure way. The MerkleTree class has methods for adding transactions to the tree, generating the root hash, and verifying the inclusion of a transaction in the tree using its hash. This implementation of the Merkle tree will be useful for ensuring the integrity of transactions in our blockchain and efficiently verifying their inclusion in the chain.